fix(web): remove unused model trigger
