--- web/app/components/header/account-setting/model-provider-page/model-parameter-modal/derive-trigger-status.ts
@@ -1,9 +1,0 @@
-import {
-  DERIVED_MODEL_STATUS_BADGE_I18N,
-  DERIVED_MODEL_STATUS_TOOLTIP_I18N,
-  deriveModelStatus,
-} from '../derive-model-status'
-
-export const deriveTriggerStatus = deriveModelStatus
-export const TRIGGER_STATUS_BADGE_I18N = DERIVED_MODEL_STATUS_BADGE_I18N
-export const TRIGGER_STATUS_TOOLTIP_I18N = DERIVED_MODEL_STATUS_TOOLTIP_I18N
