+++ web/app/components/header/account-setting/model-provider-page/model-parameter-modal/types.ts
@@ -1,0 +1,13 @@
+import type {
+  Model,
+  ModelItem,
+  ModelProvider,
+} from '../declarations'
+
+export type TriggerProps = {
+  open?: boolean
+  currentProvider?: ModelProvider | Model
+  currentModel?: ModelItem
+  providerName?: string
+  modelId?: string
+}
